==> Script/02_analyses/04a_predictive_performance_dependent.qmd <==
---
title: "Predictive Performance"
format: html
---

# Preparations

## Source Functions

To import all custom functions, we need to source the `functions.R` file. This file contains all the functions used in this project.

```{r}
source("../01_functions/functions.R")
```

## Load Packages

Additionally, we need to load the required packages.

```{r}
library(remify)
library(remstats)
library(remstimate)
library(shrinkem)
library(tidyverse)
library(MASS)
library(future)
library(future.apply)
library(patchwork)
library(knitr)
library(kableExtra)
```

## Load True Parameters and Covariates

Next, we load the true parameters and covariates used in the simulation. These are stored in the `covar.RData` and `parameters.RData` files, respectively.

```{r}
load("../../Data/01_simulation/covar.RData")
load("../../Data/01_simulation/parameters.RData")
```

## Load and Combine edgelists

Lastly, we load and combine the edgelists generated in `01_generating_data.qmd`.

```{r}
edgelists <- list()
files <- list.files("../../Data/01_simulation/01_edgelists_dependent")

# subset files with edgelist_
files <- files[grep("edgelist_", files)]


for (i in seq_along(files)) {
  load(paste0("../../Data/01_simulation/01_edgelists_dependent/", files[i]))
  
  edgelists[[i]] <- edgelist
}
```

# Compute Statistics

```{r}
calculate_statistics_parallel(edgelists = edgelists,
                              parameters = parameters,
                              covar = covar,
                              results_folder = "../../Output/01_result_files/",
                              num_cores = 20)
```

# Predictive Performance In-Sample

We first assess the predictive performance of the models in-sample. The predictive performance is calculated using the `pp_is_parallel` function, which estimates the models in parallel.

First, we do this for the full models, and then we do it for the sparse models after selecting the variables.

### Top 5%

```{r}
t1 <- Sys.time()
pp_is_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/01_top5/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               num_cores = 100,
               quantile = 0.95)
t2 <- Sys.time()
```

### Top 10%

```{r}
t1 <- Sys.time()
pp_is_parallel(edgelists = edgelists,
              m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/02_top10/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               num_cores = 100,
               quantile = 0.90)
t2 <- Sys.time()
```

### Top 20%

```{r}
t1 <- Sys.time()
pp_is_parallel(edgelists = edgelists,
              m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/03_top20/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               num_cores = 100,
               quantile = 0.80)
t2 <- Sys.time()
```

## Plots

```{r}
plot_top5 <- plot_pp(pp_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/01_top5/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 5%")

plot_top10 <- plot_pp(pp_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/02_top10/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 10%")

plot_top20 <- plot_pp(pp_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/03_top20/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 20%")
```

```{r}
# Combine the plots and collect axis titles and legends
combined_plot_pp_is <- (plot_top20 + plot_top10 + plot_top5) +
  plot_layout(guides = "collect", axis_titles = "collect") &  # Combine legends
  theme(legend.position = "bottom",
        legend.text = element_text(size = 10),
  plot.tag = element_text(size = 20, face = "bold")) & # Position the legend at the bottom
  ylim(0.65, 1)

combined_plot_pp_is <- combined_plot_pp_is +
  plot_annotation(tag_levels = "A")

ggsave("../../Output/02_plots/03_predictive_performance/pp_is_plots.png", combined_plot_pp_is, width = 12, height = 5)
```

## Sparse Models

### Top 5%

```{r}
t1 <- Sys.time()
pp_is_sparse_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/04_top5_sparse/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               selected_folder = "../../Output/01_result_files/03_selected_variables/",
               num_cores = 60,
               quantile = 0.95)
t2 <- Sys.time()
```

### Top 10%

```{r}
t1 <- Sys.time()
pp_is_sparse_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/05_top10_sparse/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               selected_folder = "../../Output/01_result_files/03_selected_variables/",
               num_cores = 100,
               quantile = 0.90)
t2 <- Sys.time()
```

### Top 20%

```{r}
t1 <- Sys.time()
pp_is_sparse_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/06_top20_sparse/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               selected_folder = "../../Output/01_result_files/03_selected_variables/",
               num_cores = 100,
               quantile = 0.80)
t2 <- Sys.time()
```

## Plots

```{r}
plot_top5 <- plot_pp_sparse(pp_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/04_top5_sparse/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 5%")

plot_top10 <- plot_pp_sparse(pp_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/05_top10_sparse/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 10%")

plot_top20 <- plot_pp_sparse(pp_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/06_top20_sparse/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 20%")
```

```{r}
# Combine the plots and collect axis titles and legends
combined_plot_pp_is <- (plot_top20 + plot_top10 + plot_top5) +
  plot_layout(guides = "collect", axis_titles = "collect") &  # Combine legends
  theme(legend.position = "bottom",
        legend.text = element_text(size = 10),
  plot.tag = element_text(size = 20, face = "bold")) & # Position the legend at the bottom
  ylim(0.6, 1)

combined_plot_pp_is <- combined_plot_pp_is +
  plot_annotation(tag_levels = "A")

ggsave("../../Output/02_plots/03_predictive_performance/pp_is_plots_sparse.png", combined_plot_pp_is, width = 12, height = 5)
```

# Out-Of Sample

The same thing is then done for the out-of-sample predictive performance. The only difference is that we use the `pp_oos_parallel` function instead of the `pp_is_parallel` function.

### Top 5%

```{r}
t1 <- Sys.time()
pp_oos_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/01_top5/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               num_cores = 100,
               quantile = 0.95,
               new = 1000)
t2 <- Sys.time()
```

### Top 10%

```{r}
t1 <- Sys.time()
pp_oos_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/02_top10/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               num_cores = 80,
               quantile = 0.90,
               new = 1000)
t2 <- Sys.time()
```

### Top 20%

```{r}
t1 <- Sys.time()
pp_oos_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/03_top20/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               num_cores = 80,
               quantile = 0.80,
               new = 1000)
t2 <- Sys.time()
```

## Plots

```{r}
plot_top5 <- plot_pp(pp_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/01_top5/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 5%")

plot_top10 <- plot_pp(pp_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/02_top10/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 10%")

plot_top20 <- plot_pp(pp_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/03_top20/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 20%")
```

```{r}
# Combine the plots and collect axis titles and legends
combined_plot_pp_oos <- (plot_top20 + plot_top10 + plot_top5) +
  plot_layout(guides = "collect", axis_titles = "collect") &  # Combine legends
  theme(legend.position = "bottom",
        legend.text = element_text(size = 10),
  plot.tag = element_text(size = 20, face = "bold")) & # Position the legend at the bottom
  ylim(0.7, 1)

combined_plot_pp_oos <- combined_plot_pp_oos +
  plot_annotation(tag_levels = "A")

ggsave("../../Output/02_plots/03_predictive_performance/pp_oos_plots.png", combined_plot_pp_oos, width = 12, height = 5)
```

## Sparse Models

### Top 5%

```{r}
t1 <- Sys.time()
pp_oos_sparse_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/04_top5_sparse/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               selected_folder = "../../Output/01_result_files/03_selected_variables/",
               num_cores = 100,
               quantile = 0.95,
               new = 1000)
t2 <- Sys.time()
```

### Top 10%

```{r}
t1 <- Sys.time()
pp_oos_sparse_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/05_top10_sparse/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               selected_folder = "../../Output/01_result_files/03_selected_variables/",
               num_cores = 100,
               quantile = 0.90,
               new = 1000)
t2 <- Sys.time()
```

### Top 20%

```{r}
t1 <- Sys.time()
pp_oos_sparse_parallel(edgelists = edgelists,
               m = c(100, 200, 400, 800, 1600, 3200, 6400), 
               parameters,
               covar,
               results_folder = "../../Output/01_result_files/",
               output_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/06_top20_sparse/",
               statistics_folder = "../../Output/01_result_files/06_statistics/",
               selected_folder = "../../Output/01_result_files/03_selected_variables/",
               num_cores = 100,
               quantile = 0.80,
               new = 1000)
t2 <- Sys.time()
```

## Plots

```{r}
plot_top5 <- plot_pp_sparse(pp_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/04_top5_sparse/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 5%")

plot_top10 <- plot_pp_sparse(pp_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/05_top10_sparse/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 10%")

plot_top20 <- plot_pp_sparse(pp_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/06_top20_sparse/",
        m = c(100, 200, 400, 800, 1600, 3200, 6400),
        num_edgelists = 100,
        title = "Top 20%")
```

```{r}
# Combine the plots and collect axis titles and legends
combined_plot_pp_oos <- (plot_top20 + plot_top10 + plot_top5) +
  plot_layout(guides = "collect", axis_titles = "collect") &  # Combine legends
  theme(legend.position = "bottom",
        legend.text = element_text(size = 10),
  plot.tag = element_text(size = 20, face = "bold")) & # Position the legend at the bottom
  ylim(0.7, 1)

combined_plot_pp_oos <- combined_plot_pp_oos +
  plot_annotation(tag_levels = "A")

ggsave("../../Output/02_plots/03_predictive_performance/pp_oos_plots_sparse.png", combined_plot_pp_oos, width = 12, height = 5)
```


# Tables

## Full Models

```{r}
table_is_full <- table_pp(pp_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/",
         subfolders = c("01_top5", "02_top10", "03_top20"),
         m = c(100, 200, 400, 800, 1600, 3200, 6400),
         num_edgelists = 100)$latex

save(table_is_full, file= "../../Output/03_tables/table_pp_is_full_dependent.tex")

table_oos_full <- table_pp(pp_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/",
         subfolders = c("01_top5", "02_top10", "03_top20"),
         m = c(100, 200, 400, 800, 1600, 3200, 6400),
         num_edgelists = 100)$latex

save(table_oos_full, file= "../../Output/03_tables/table_pp_oos_full_dependent.tex")
```

## Sparse Models

```{r}
table_is_sparse <- table_pp_sparse(pp_folder = "../../Output/01_result_files/04a_predictive_performance_is_dependent/",
         subfolders = c("04_top5_sparse", "05_top10_sparse", "06_top20_sparse"),
         m = c(100, 200, 400, 800, 1600, 3200, 6400),
         num_edgelists = 100)$latex

save(table_is_sparse, file= "../../Output/03_tables/table_pp_is_sparse_dependent.tex")

table_oos_sparse <- table_pp_sparse(pp_folder = "../../Output/01_result_files/04b_predictive_performance_oos_dependent/",
         subfolders = c("04_top5_sparse", "05_top10_sparse", "06_top20_sparse"),
         m = c(100, 200, 400, 800, 1600, 3200, 6400),
         num_edgelists = 100)$latex

save(table_oos_sparse, file= "../../Output/03_tables/table_oos_sparse_dependent.tex")
```
